+++ menu/savestateio.h
@@ -4,7 +4,7 @@
 void SetSaveStateIoModeMemory(unsigned char **data);
 void SetSaveStateIoModeFile();
 
-int  statef_open(const char *fname, const char *mode);
+long  statef_open(const char *fname, const char *mode);
 int  statef_read(void *p, int l);
 int  statef_write(void *p, int l);
 void statef_close();
